add findAll present student after finish the game
